requirements.txt
fastapi==0.110.0
pandas==2.2.1
openai==1.12.0
uvicorn==0.27.1
llama-index==0.10.12
Jinja2==3.1.3
python-dotenv==1.0.1